完成！

+++ work08/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from .models import Memo
 from django.shortcuts import get_object_or_404
+from django.http import HttpResponseBadRequest
 
 
 def top(request):
@@ -26,13 +27,21 @@ def edit(request, memo_id):
     return render(request, "work08/edit.html", {"memo": memo})
 
 
+# 新規作成用のビュー関数
 def create_memo(request):
-    memo = Memo.objects.create(title="no title", content="")
-    return redirect("work08:edit", memo_id=memo.id)
+    if request.method == "POST":
+        memo = Memo.objects.create(title="no title", content="")
+        return redirect("work08:edit", memo_id=memo.id)
+    else:
+        return HttpResponseBadRequest("無効なリクエストです。")
 
 
 # 削除用のビュー関数
 def delete_memo(request, memo_id):
-    memo = get_object_or_404(Memo, id=memo_id)
-    memo.delete()  # データベースから削除
-    return redirect("work08:top")  # 一覧画面に戻る
+    if request.method == "POST":
+        memo = get_object_or_404(Memo, id=memo_id)
+        memo.delete()
+        return redirect("work08:top")
+    else:
+        # 不正アクセス防止（GETで来たらエラーを返す）
+        return HttpResponseBadRequest("無効なリクエストです。")
